big cleanup on sound effect enums, add ovl62 without data

+++ src/ovl7.c
@@ -1423,7 +1423,7 @@ void scManager_TrainingMode_InitScene()
 
 	func_80007024(&D_ovl7_80190870);
 
-	D_ovl7_8019088C.arena_size = ((uintptr_t)&lOverlay7ArenaHi - (uintptr_t)&lOverlay7ArenaLo);
+	D_ovl7_8019088C.arena_size = (size_t) ((uintptr_t)&lOverlay7ArenaHi - (uintptr_t)&lOverlay7ArenaLo);
 	D_ovl7_8019088C.proc_start = scTrainingMode_InitTrainingMode;
 
 	do
